update_install: Install packages from each patch separately
Each patch and involved packages, in the single patch, are installed and
patched. This will result in less conflicts between packages.
Added record info for overview what is happening.

Most of the code is just moved into loop which will do the whole preinstall,
patch and verify process. New hashes and arrays are created within the loop
containing only packages/binaries of the patch.

Update repos in the loop have to be disabled and enabled, added in following
615ec7c804d65a652ae3f477e0efd03243ca7069